repo cleaning & restructure
